requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
langdetect==1.0.9
googletrans==4.0.0rc1
PyMuPDF==1.23.8
python-docx==1.1.0
fpdf2==2.7.6
aiofiles==23.2.1 